requirements.txt
altgraph==0.17.3
anyio==3.6.2
certifi==2022.12.7
colorama==0.4.6
flet==0.4.2
flet-core==0.4.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
loguru==0.6.0
oauthlib==3.2.2
packaging==23.0
pefile==2023.2.7
psutil==5.9.4
pyinstaller==5.8.0
pyinstaller-hooks-contrib==2023.0
python-reapy==0.10.0
pywin32-ctypes==0.2.0
repath==0.9.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
typing-extensions==4.5.0
watchdog==2.3.0
websocket-client==1.5.1
websockets==10.4
win32-setctime==1.1.0
